server/requirements.txt
flask
sendgrid
pillow
dotenv
requests
groq
openpyxl